+++ source/config/defaults.js
@@ -1,6 +1,7 @@
 import chars from '../data/common-chars.json' with { type: 'json' }
 import htmle from '../data/html-entities.json' with { type: 'json' }
-import s from '../lib/reparts.js'
+import * as entities from '../entities/index.js'
+import s from '#lib/reparts.js'
 
 
 export default
@@ -52,50 +53,11 @@ export default
         '>',
     ],
 
-    delouse:
-    {
-        autolink:
-        {
-            value: [ 'htmlEntsToChars', 'commonHtmlEnts' ],
-            url: [ 'htmlEntsToChars', 'encodeUriChars', 'commonHtmlEnts' ]
-        },
-        code: [ 'lineEndToSpace', 'trimEncSpace', 'commonHtmlEnts' ],
-        codeblock: [ 'commonHtmlEnts' ],
-        fenceblock:
-        {
-            value: [ 'commonHtmlEnts' ],
-            info: [ 'common' ]
-        },
-        email:
-        {
-            value: [ 'commonHtmlEnts' ],
-            email: [ 'commonHtmlEnts' ]
-        },
-        emphasis: [ 'common' ],
-        header: [ 'common' ],
-        image:
-        {
-            value: [ 'common' ],
-            href: [ 'uri' ],
-            title: [ 'common' ]
-        },
-        link:
-        {
-            value: [ 'common' ],
-            href: [ 'uri' ],
-            title: [ 'common' ]
-        },
-        paragraph: [ 'trimAroundNewline', 'common' ],
-        quotation: [ 'trimEnd' ],
-        setext: [ 'trimAroundNewline', 'common' ],
-        strong: [ 'common' ],
-    },
-
     delousers:
     {
         ampersandToEnt: { search: [ `${s.ne}&(?!(amp|quot|gt|lt);)`, 'g' ], replace: '&amp;' },
         commonCharToEnt: { search: [ '(?<cap>[<>"])', 'g' ], replace: g => chars[g.cap] },
-        unescapePunct: { search: [ `\\\\(?<cap>[${s.apc}])`, 'g' ], replace: '{cap}' },
+        unescapePunct: { search: [ `\\\\(?<cap>[${s.apc}])`, 'g' ], replace: '$<cap>' },
         namedEntToChar: 
         { 
             search: [ `${s.ne}(?<cap>&[a-zA-Z0-9]+?;)`, 'g' ], 
@@ -114,7 +76,7 @@ export default
         lineEndToSpace: { search: [ `[${s.le}]`, 'g' ], replace: ' ' },
         encodeUriChars: { search: [ '(?<cap>[^%\\w]+)', 'g' ], replace: g => encodeURI(g.cap) },
         trimEnd: { search: '\\s*\\n$', replace: '' },
-        trimEncSpace: { search: [ '^ (?<cap>\\s*[^\\s].*) $' ], replace: '{cap}' },
+        trimEncSpace: { search: [ '^ (?<cap>\\s*[^\\s].*) $' ], replace: '$<cap>' },
         trimAroundNewline: { search: [ '\\s*\\n\\s*', 'g' ], replace: '\n' },
         // combo-delousers
         common: [ 'htmlEntsToChars', 'unescapePunct', 'commonHtmlEnts' ],
@@ -122,6 +84,8 @@ export default
         htmlEntsToChars: [ 'namedEntToChar', 'decimalEntToChar', 'hexEntToChar' ],
         uri: [ 'htmlEntsToChars', 'unescapePunct', 'encodeUriChars', 'commonHtmlEnts' ]
     },
+
+    entities,
 
     fm:
     {
@@ -136,37 +100,6 @@ export default
     {
         variables: /\{([\w.]+)(?:\?\?(.*?))?\}/g,
         segments: /\{\?(((?!\{\?).)+?)\?\}/g
-    },
-
-    nestable:
-    {
-        emphasis: [ 'autolink', 'code', 'email', 'emphasis', 'image', 'linebreak', 'link' ],
-        header: [ 'autolink', 'code', 'email', 'emphasis', 'html', 'link' ],    
-        image: [ 'autolink', 'code', 'email', 'emphasis', 'html', 'image', 'link' ],
-        link: [ 'autolink', 'code', 'email', 'emphasis', 'html', 'image', 'link' ],
-        listitem:
-        [ 
-            'arbitag', 'autolink', 'code', 'codeblock', 'codefence', 
-            'divide', 'email', 'emphasis', 'header', 'html',
-            'htmlblock', 'image', 'linebreak', 'link', 'list',
-            'paragraph', 'quotation', 'reference', 'setext'
-        ],
-        paragraph: [ 'autolink', 'code', 'email', 'emphasis', 'html', 'image', 'linebreak', 'link' ],
-        quotation:
-        [ 
-            'codeblock', 'codefence', 
-            'divide', 'header',
-            'htmlblock', 'list',
-            'paragraph', 'quotation', 'reference', 'setext',
-        ],
-        root:
-        [ 
-            'arbitag', 'autolink', 'code', 'codeblock', 'codefence', 
-            'divide', 'email', 'emphasis', 'header', 'html',
-            'htmlblock', 'image', 'linebreak', 'link', 'list',
-            'paragraph', 'quotation', 'reference', 'setext',
-        ],
-        setext: [ 'autolink', 'code', 'emphasis', 'html', 'image', 'linebreak', 'link' ]
     },
 
     onAction: null,