+++ source/entities/codeblock.js
@@ -1,0 +1,19 @@
+
+/*
+    Indented code block.
+*/
+export default
+{
+    order: 30,
+
+    pattern: 'block-indented-content',
+    patternData:
+    {
+        rejectOnForcedClose: true
+    },
+
+    delouse:
+    {
+        value: [ 'commonHtmlEnts' ]
+    }   
+}