+++ source/entities/divide.js
@@ -1,0 +1,11 @@
+
+/*
+    Thematic break.
+*/
+export default
+{
+    order: 20,
+
+    pattern : 'block-char-divider',
+    patternData: { chars: '-_*' }
+}